Extract addMisstrokeStatus.ts
